+++ semana24/numero-faltante/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "knex-express",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "start": "tsc && node ./build/index.js",
+    "test": "jest"
+  },
+  "author": "Dienay",
+  "license": "ISC",
+  "dependencies": {
+    "@types/jest": "^26.0.15",
+    "@types/node": "13.7.7",
+    "jest": "^26.6.2",
+    "ts-jest": "^26.4.3"
+  }
+}
