+++ semana24/numero-faltante/src/index.test.ts
@@ -1,0 +1,8 @@
+import { entrada1, numeroFaltante } from ".";
+
+describe("Testando Exercício", ()=>{
+    test("Testando o encontrar valor retornado", ()=>{
+        const result = numeroFaltante(entrada1)
+        expect(result).toBe(47)
+    });
+})
